Refactoring a lot of test support code so it isn't spread all over
